improve: CMDB插件多版本目录改造 #2052
