+++ pipeline_plugins/components/collections/sites/open/cc/replace_fault_machine/__init__.py
@@ -10,3 +10,5 @@ Unless required by applicable law or agreed to in writing, software distributed 
 an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
+
+from .legacy import CCReplaceFaultMachineComponent  # noqa
