+++ pipeline_plugins/components/static/components/atoms/cc/create_set/legacy.js
