+++ pipeline_plugins/components/static/components/atoms/cc/create_set/v1_0.js
